add ArmA 2 Operation Arrowhead client
This client supports Operation Arrowhead and was tested on the
United Operations primary server.

This client subclasses GenericGamespyClient and adds the following methods:

    class.parse_response()
    class.data()

The former returns None and just handles the parsing of the data.
The latter returns a dictionary of the data from the server. If the
internal dict is not populated, data() calls parse_response().

It is recommended that both of these methods be implemented in any
subclasses of GenericGamespyClient.